finished separating out the screens into separate files, will need to go through them now to clean them up for efficiency and style
